requirements.txt
click==7.1.2
Flask==1.1.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
#pkg-resources>=0.0.0
PyYAML==6.0
Werkzeug==1.0.1
